requirements.txt
oura-ring>=2.0.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.12.0
scikit-learn>=1.1.0
python-dotenv>=0.19.0
requests>=2.28.0

